requirements.txt
opencv-python==4.8.1.78
Pyrogram==2.0.106
numpy==1.26.4
opencv-python==4.8.1.78
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
numpy==1.26.4
opencv-python==4.8.1.78
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
TgCrypto==1.2.5
install==1.3.5
numpy==1.26.4
opencv-python==4.8.1.78
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.1
TgCrypto==1.2.5
